Make deletion of files optional in update command

+++ chef_solo_cup/commands/update.py
@@ -5,13 +5,14 @@ from chef_solo_cup.commands.sync import sync
 from chef_solo_cup.log import setup_custom_logger
 
 
-def update(args, config, logger=None):
+def update(args, config, delete_files=False, logger=None):
     if logger is None:
         logger = setup_custom_logger('chef-solo-cup', args)
 
     sync(args, config, logger=logger)
 
     command = "cd {0} && source /etc/profile && `which chef-solo` -c {0}/{1} -j {0}/{2} -l {3}"
-    command += ' && rm -rf {0}'.format(args.chef_file_dest)
+    if delete_files:
+        command += ' && rm -rf {0}'.format(args.chef_file_dest)
 
     return sudo_dry(command.format(args.chef_file_dest, args.config_path, config.get('dna_path'), args.loglevel), args, logger=logger)
